+++ src/lib.rs
@@ -1,6 +1,9 @@
 mod bindings;
 pub mod client;
 mod interop;
+mod checkout_component;
+
 
 pub use bindings::*;
 pub use interop::*;
+pub use checkout_component::*;